requirements.txt
Flask
flask-cors
pymongo
dnspython
requests
beautifulsoup4
python-dotenv







